requirements.dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.dev.in
#
-e file:.
    # via -r requirements.dev.in
appdirs==1.4.4
    # via black
attrs==21.2.0
    # via
    #   jsonschema
    #   pytest
black==21.5b2
    # via -r requirements.dev.in
click==8.0.1
    # via black
flake8==3.9.2
    # via -r requirements.dev.in
iniconfig==1.1.1
    # via pytest
isort==5.8.0
    # via -r requirements.dev.in
jsonschema==3.2.0
    # via pipeline
mccabe==0.6.1
    # via flake8
mypy-extensions==0.4.3
    # via black
packaging==20.9
    # via pytest
pathspec==0.8.1
    # via black
pluggy==0.13.1
    # via pytest
py==1.10.0
    # via pytest
pycodestyle==2.7.0
    # via flake8
pyflakes==2.3.1
    # via flake8
pyparsing==2.4.7
    # via packaging
pyrsistent==0.17.3
    # via jsonschema
pytest==6.2.4
    # via -r requirements.dev.in
regex==2021.4.4
    # via black
ruamel.yaml.clib==0.2.2
    # via ruamel.yaml
ruamel.yaml==0.17.7
    # via pipeline
six==1.16.0
    # via jsonschema
toml==0.10.2
    # via
    #   black
    #   pytest

# The following packages are considered to be unsafe in a requirements file:
# setuptools
